requirements.txt
keras==3.0.5
numpy==2.0.1
opencv_python==4.10.0.82
opencv_python_headless==4.10.0.82
Pillow==10.4.0
spotipy==2.24.0
